Server/requirements.txt
fastapi
uvicorn[standard]
jinja2
paho-mqtt
python-dotenv
